Добавлена страница для админа, изменена кнопка вход/выход
