+++ app/templates/admin/index.html
@@ -1,10 +1,9 @@
 {% extends 'base.html' %}
 
 {% block title %}
-{{ page_title }}
+{{page_title}}
 {% endblock %}
 
-
 {% block content %}
-{{ content }}
+  <h1>{{ content }}</h1>
 {% endblock %}
