+++ app/templates/base.html
@@ -4,67 +4,79 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>{% block title %}{% endblock %}</title>
-
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}">
   </head>
-  <body class="d-flex flex-column h-100">
 
-<div class="page">
+<body class="d-flex flex-column min-vh-100">
 
-    <header class="d-flex flex-wrap align-items-center justify-content-center justify-content-md-between py-3 mb-4 border-bottom">
-      <div class="col-md-3 mb-2 mb-md-0">
-        <a href="/" class="d-inline-flex link-body-emphasis text-decoration-none">
-          <img class="header__img" height="48" width="48" src="{{ url_for('static', filename='img/logo.png') }}" alt="logo">
-        </a>
-      </div>
-
-      <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-        <li><a href="{{ url_for('index') }}" class="nav-link px-2 link-secondary">Главная</a></li>
-        <li><a href="{{ url_for('shop') }}" class="nav-link px-2">Магазин</a></li>
-        <li><a href="{{ url_for('contacts') }}" class="nav-link px-2">Контакты</a></li>
-        <li><a href="{{ url_for('questions') }}" class="nav-link px-2">FAQs</a></li>
-        <li><a href="{{ url_for('about') }}" class="nav-link px-2">О нас</a></li>
-      </ul>
-
-      <div class="col-md-3 text-end">
-        <a class="btn btn-outline-primary" href="{{ url_for('login') }}" role="button">Войти</a>
-
-        <button type="button" class="btn btn-primary">Зарегистрироваться</button>
-      </div>
-    </header>
-
-  <div class="content">
-        {% with messages = get_flashed_messages() %}
-      {% if messages %}
-        <div class="alert alert-warning" role="alert">
-          {% for message in messages %}
-            {{ message }}<br>
-          {% endfor %}
-        </div>
-      {% endif %}
-    {% endwith %}
-      {% block content %}{% endblock %}
-  </div>
-
-  <footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top">
-    <div class="col-md-4 d-flex align-items-center">
-      <a href="/" class="mb-3 me-2 mb-md-0 text-body-secondary text-decoration-none lh-1">
-        <svg class="bi" width="30" height="24"><use xlink:href="#bootstrap"></use></svg>
+<header class="fixed-top bg-white shadow-sm">
+  <div class="container">
+    <div class = "d-flex flex-wrap align-items-center justify-content-center justify-content-md-between py-3 mb-3 border-bottom">
+    <div class="col-md-3 mb-2 mb-md-0">
+      <a href="/" class="d-inline-flex link-body-emphasis text-decoration-none">
+        <img class="header__img" height="48" width="48" src="{{ url_for('static', filename='img/logo.png') }}" alt="logo">
       </a>
-      <span class="mb-3 mb-md-0 text-body-secondary">© 2024 Company, Inc</span>
     </div>
 
-    <ul class="nav col-md-4 justify-content-end list-unstyled d-flex">
-      <li class="ms-3"><a class="text-body-secondary" href="https://vk.com/il1uminaterrr"><img class="header__img" height="30" width="30" src="{{ url_for('static', filename='img/vk.png') }}" alt="VK"></a></li>
-      <li class="ms-3"><a class="text-body-secondary" href="/"><img class="header__img" height="30" width="30" src="{{ url_for('static', filename='img/telegram.png') }}" alt="low"></a></li>
-      <li class="ms-3"><a class="text-body-secondary" href="/"><img class="header__img" height="30" width="30" src="{{ url_for('static', filename='img/inst.png') }}" alt="low"></a></li>
+    <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
+      <li><a href="{{ url_for('/.index') }}" class="nav-link px-2 link-secondary">Главная</a></li>
+      <li><a href="{{ url_for('shop.shop') }}" class="nav-link px-2">Магазин</a></li>
+      <li><a href="{{ url_for('/.contacts') }}" class="nav-link px-2">Контакты</a></li>
+      <li><a href="{{ url_for('/.questions') }}" class="nav-link px-2">FAQs</a></li>
+      <li><a href="{{ url_for('/.about') }}" class="nav-link px-2">О нас</a></li>
     </ul>
-  </footer>
-</div>
+      <div class="col-md-3 text-end">
+          {% if current_user.is_anonymous %}
+          <a class="btn btn-outline-primary" href="{{ url_for('user.login') }}" role="button">Войти</a>
+          {% else %}
+          <a class="btn btn-outline-primary" href="{{ url_for('user.logout') }}" role="button">Выйти</a>
+          {% endif %}
+        <button type="button" class="btn btn-primary">Зарегистрироваться</button>
+        </div>
+    </div>
+    </div>
+</header>
 
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js" integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js" integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy" crossorigin="anonymous"></script>
-  </body>
+<main class="container mt-5 " style="padding-top: 80px;">
+  <div class="row justify-content-center align-items-center">
+    {% with messages = get_flashed_messages() %}
+    {% if messages %}
+      <div class="alert alert-warning" role="alert">
+        {% for message in messages %}
+          {{ message }}<br>
+        {% endfor %}
+      </div>
+        {% endif %}
+        {% endwith %}
+
+    {% block content %}{% endblock %}
+  </div>
+</main>
+
+<footer class="bg-white border-top mt-auto">
+  <div class="container">
+    <div class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top">
+      <div class="col-md-4 d-flex align-items-center">
+        <a href="/" class="d-inline-flex link-body-emphasis text-decoration-none">
+            <img class="header__img" height="30" width="30" src="{{ url_for('static', filename='img/logo.png') }}" alt="logo">
+          </a>
+        <span class="mb-3 mb-md-0 text-body-secondary">  © 2024 Company, Inc</span>
+      </div>
+
+      <ul class="nav col-md-4 justify-content-end list-unstyled d-flex">
+        <li class="ms-3"><a class="text-body-secondary" href="https://vk.com/il1uminaterrr"><img class="header__img" height="30" width="30" src="{{ url_for('static', filename='img/vk.png') }}" alt="VK"></a></li>
+        <li class="ms-3"><a class="text-body-secondary" href="https://t.me/Il1uminater"><img class="header__img" height="30" width="30" src="{{ url_for('static', filename='img/telegram.png') }}" alt="TG"></a></li>
+        <li class="ms-3"><a class="text-body-secondary" href="/"><img class="header__img" height="30" width="30" src="{{ url_for('static', filename='img/inst.png') }}" alt="low"></a></li>
+      </ul>
+    </div>
+  </div>
+</footer>
+
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
+  <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js" integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r" crossorigin="anonymous"></script>
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js" integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy" crossorigin="anonymous"></script>
+</body>
 </html>
+
+
