+++ app/templates/main/about.html
@@ -7,4 +7,5 @@
 
 {% block content %}
 {{ content }}
+
 {% endblock %}
