+++ app/templates/main/index.html
@@ -5,6 +5,7 @@
 {% endblock %}
 
 
-{% block content %}   
-{{ content }}        
+{% block content %}
+{{ content }}
+
 {% endblock %}
