+++ app/templates/page_error.html
@@ -181,7 +181,7 @@ a:hover {
           <img src="{{ url_for('static', filename='img/error-404.png') }}" alt="Страница не найдена (ошибка 404)">
           <h2><b>404</b> Страница не найдена</h2>
           <p>посетите главную страницу <br> возможно вы найдёте её</p>
-          <a href="{{ url_for('index') }}" class="cmn-btn mt-4">На главную</a>
+          <a href="{{ url_for('/.index') }}" class="cmn-btn mt-4">На главную</a>
         </div>
       </div>
     </div>
