+++ app/templates/user/login.html
@@ -4,22 +4,24 @@
 {{ page_title }}
 {% endblock %}
 
-{% block content %}
 
-<div class="row justify-content-center mt-5">
-  <div class="col-md-6 col-lg-12">
-  <form action="{{ url_for('process_login') }}" method="POST" class="border p-4 rounded shadow-sm">
+
+{% block content %}
+<div class="container">
+  <div class="row justify-content-center align-items-center">
+  <div class="col-12 col-sm-10 col-md-8 col-lg-6 col-xl-4">
+  <form action="{{ url_for('user.process_login') }}" method="POST" class="border p-4 rounded shadow-sm">
     <img class="mb-3 mx-auto d-block" src="{{ url_for('static', filename='img/logo.png') }}" alt="" width="72" height="57">
       {{ form.hidden_tag() }}
-      <div class="form-group mb-3">
+      <div class="form-group mb-4">
         {{ form.username.label }}
         {{ form.username() }}
       </div>
-      <div class="form-group mb-3">
+      <div class="form-group mb-4">
         {{ form.password.label }}
         {{ form.password() }}
       </div>
-    <div class="form-group form-check">
+    <div class="form-group form-check mb-4">
       {{ form.remember_me() }}
       {{ form.remember_me.label(class_='form-check-label') }}
     </div>
@@ -28,6 +30,10 @@
     </div>
   </form>
       </div>
+  </div>
 </div>
 {% endblock %}
+
+
+
