+++ create_admin.py
@@ -2,19 +2,19 @@ from getpass import getpass
 import sys
 
 from app import create_app
-from app.models import User, db
+from app.user.models import User, db
 
 app = create_app()
 
 with app.app_context():
     email = input('Введите email: ')
 
-    if User.query.filter(User.email == email).count():
+    if User.query.filter(User.email.ilike(email)).count():
         print('Этот Email уже используется')
         sys.exit(0)
 
     username = input('Введите имя: ')
-    if User.query.filter(User.username == username).count():
+    if User.query.filter(User.username.ilike(username)).count():
         print('Имя занято')
         sys.exit(0)
     password1 = getpass('Введите пароль')
